+++ src/app/app.component.ts
@@ -1,4 +1,4 @@
-import { Component } from '@angular/core';
+import { Component, ViewChild, ElementRef, AfterViewInit } from '@angular/core';
 import { CartService } from './core/cart.service';
 
 @Component({
@@ -6,8 +6,12 @@ import { CartService } from './core/cart.service';
   templateUrl: './app.component.html',
   styleUrls: ['./app.component.css']
 })
-export class AppComponent {
-  title = 'shop';
-  
+export class AppComponent implements AfterViewInit {
+  @ViewChild('appTitle') title: ElementRef;
+
   constructor(public cartServise: CartService) {}
+
+  ngAfterViewInit() {
+    this.title.nativeElement.innerHTML = 'Find your best sport shoes';
+  }
 }