+++ src/app/app.module.ts
@@ -2,23 +2,20 @@ import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 
 import { AppComponent } from './app.component';
-import { FirstComponent } from './first/first.component';
-import { ButtonComponent } from './shared/button/button.component';
-import { ProductsListComponent } from './products-list/products-list.component';
-import { ProductComponent } from './products-list/product/product.component';
-import { CartListComponent } from './cart-list/cart-list.component';
+
+import { SharedModule } from './shared/shared.module';
+import { CartModule } from './cart-list/cart.module';
+import { ProductModule } from './products/product.module';
 
 @NgModule({
   declarations: [
     AppComponent,
-    FirstComponent,
-    ButtonComponent,
-    ProductsListComponent,
-    ProductComponent,
-    CartListComponent
   ],
   imports: [
-    BrowserModule
+    BrowserModule,
+    SharedModule,
+    CartModule,
+    ProductModule
   ],
   providers: [],
   bootstrap: [AppComponent]