+++ src/app/core/cart.service.ts
@@ -15,24 +15,49 @@ export class CartService {
     return this.basketItems;
   }
 
-  addItemtoCart(item: Product): void {
+  addItemtoCart(data: any): void {
     this.newCartItem = {
-      photo: item.photo, 
-      name: item.name, 
-      price: item.price, 
+      photo: data.item.photo,
+      name: data.item.name,
+      price: data.item.price,
+      size: data.size,
       quantity: 1
-    }
+    };
 
     this.basketItems.forEach(x => {
-      if(x.name === item.name) {
+      if (x.name === data.item.name && x.size === data.size) {
         x.quantity += 1;
-        delete this.newCartItem
+        this.newCartItem = null;
       }
     });
 
-    if(this.newCartItem) {
+    if (this.newCartItem) {
       this.basketItems.push(this.newCartItem);
-    }    
+    }
+  }
+
+  reduceItemQty(item: CartItem): void {
+    this.basketItems.forEach(currentItem => {
+      if (currentItem.name === item.name && currentItem.size === item.size) {
+        currentItem.quantity -= 1;
+      }
+
+      if (currentItem.quantity === 0) {
+        this.removeItem(currentItem);
+      }
+    });
+  }
+
+  increaseItemQty(item: CartItem): void {
+    this.basketItems.forEach(currentItem => {
+      if (currentItem.name === item.name && currentItem.size === item.size) {
+        currentItem.quantity += 1;
+      }
+    });
+  }
+
+  removeItem(item: CartItem): void {
+    this.basketItems = this.basketItems.filter(currentItem => currentItem !== item);
   }
 
 }