+++ src/app/products/choose-size.model.ts
@@ -1,0 +1,4 @@
+export interface ChooseSize {
+    size: number;
+    active: boolean;
+}