+++ src/app/shared/product.model.ts
@@ -1,5 +1,5 @@
 
-import { Category } from './';
+import { Category } from './category.enum';
 
 export interface Product {
     name: string;
@@ -7,6 +7,7 @@ export interface Product {
     price: number;
     category: Category;
     isAvailable: boolean;
+    quantity?: Array<number>;
     sizes?: Array<number>;
     color: string;
     photo: string;
@@ -16,5 +17,6 @@ export interface CartItem {
     photo: string;
     name: string;
     price: number;
+    size: number;
     quantity: number;
 }